Add therapist and client sessions
This commit does a lot. In the future I'll split things up more
now that the basic MVP is done.

I split up the JS into a few different prototypes. `Lightbar`
should strictly be used for controlling the behavior of a
rendered Lightbar. This is the bit used by both the therapist
and the client so that the therapist and the client see the same
thing on their screen.

SessionControls modifies the state of the therapist's lightbar
and also emits a new settings event. This event is broadcast
on the therapist's current sessions namespace.

ClientSession connects the person receiving EMDR therapy to their
therapist's lightbar session. All this does is set up to
receive new settings whenever the therapist changes the lightbar
settings (including starting and stopping the lightbar movement).

Session IDs are currently just three letters and three numbers.
I like this for now because it's pretty easy to send the link
as a text message or to dictate the link over the phone. I also
prefer it to allowing the therapist to set their own session ID
because it removes all risk for a therapist to accidentally include
PPI, which we are explicitly chosing not to handle in this app.

I did everything in kind of an older JS style. Mostly just because
I wanted to see if I remembered how prototypes worked. There's
nothing super interesting in this app. The animation stuff is simple.
There may be room for EE in the animation, there's definitely some
duplication in the _moveDirection functions... but my gut right now
says that trying to abstract that will end up making the code less clear.