Use percentage for width
Using a percentage means that irrespective of the width of the
therapist and the patient's screens (and notably the difference
between the two), each will still see a light and lightbar which
have the same proportions as the other. This in turn also makes
the speed across both all screens consistent, as the speed is
taken as a "length for one bounce" and because the size of the
light on both screens shares the same proportions, the light's
size will compensate for what speed would have left it lagging.

It is easy to prove this by testing with a very small viewport
and a very large viewport, ensuring their settings are synced
(the easiest way is through a refresh) and starting them at the
same time. You will observe that despite the larger viewport
have a farther distance for the light to travel, they both arrive
at each end at the same time.

I think this is actually not ideal because the therapist does
not necessarily need to control the "bounces per minute", which
is now what the speed will correspond to. Rather I think the speed
should correspond actually to the rate of change, meaning that
on each screen, irrespective as to the size of the light and
viewport, each light will be traveling at the same speed, i.e.,
covering the same number of pixels over the same period of time.