Sync therapist settings on page load
If the therapist has previously configured the lightbar, some
browsers will store the position of the width and speed ranges.
This was causing the width and speed of the light to not
reflect the actual settings of the ranges until the therapist
made a change to both (in which case only the one that was
changed was taken into account and reset to the actual value
being displayed, the new value). If we evaluate the value
of each range input when the controls are initialized, then we
can go ahead and make sure that the settings are synchronized
with the ranges themselves. Naturally we will also emit these
updated settings to the client.