Add docker-compose with four workers
Runs 4 workers each with uWSGI in front of gevent. nginx is used
to load balance between the four.

Future work might look into using docker swarm to spin up workers
instead of hard coding 4 of them, but probably not worth the effort
until far into the future anyway. For now the current hard-coded
version should work okay.

Redis is used to allow each of the workers to communicate about
connected sockets and dispatch when broadcast is used. We do use
broadcast to connect therapist events to client events so this is
necessary. In any case, I'm sure we'll find another use for Redis.

We might, for example, start storing the session in redis and use
Flask-Session rather than the cookie-based session storage that
is currently used. Not sure if there's a huge win there but it's
something to keep in mind going forward.