Use html templates instead of static rendering
This allows us to get rid of the "therapist-init-session" event
that we were emitting when the therapist connected to get the
session id. This also reduces some duplicated code in the
templates and generally does things in a more conventional flask
way (less MVP, if you will).

I also renamed client to patient becasue "client" means too many
things in computer/internet land and patient is clearer anyway.