Add audio version
We're duplicating some patterns here, particularly around the audio/lightbar
settings, the template, the controls, and especially the controllers. There
is definitely going to be some room for improvement here once the dust
settles and I can get a clearer picture of how to abstract these behaviors.

Pitch manipulation is still left, and adding a volume knob is probably going
to be a good idea.

I'm starting to wonder if pitch and volume might be best controlled by the
patient rather than the therapist. Potentially even speed. Same for the lightbar.
I'll have to ask for some input on that to get a better idea about what would
be the easiest to use but also provide the fewest distraction. A potential
solution for this is to have the starting and stopping be controllable by
the therapist but have the patient set up their preferences for the
speed, pitch, width, etc. and also have a start and stop button. We'd essentially
be syncing in the opposite direction. If we did that we could just let the
patient open the session and have the therapist connect to it and have a start/stop
button, who knows. Anyway, just something to contemplate.