Prevent popping while stopping, starting, and panning
Due to the nature of how the human ear works[1] with sine waves
to prevent popping, we have to do all sorts of juggling with
linear ramps to values over time.

Basically, because we cannot actually stop the sound on the mid-
point of the wave (the audio API doesn't provide us with the
ability to do something so precise) we just need to follow a strategy
of gradually adjusting the gain (for stop/start) in each direction
to approach zero, until the popping that will occur from the
speakers actually stopping the noise production will be imperceptibly
quiet.

The article linked below suggests using setTargetAtTime and the MDN
documentation recommends the exponential ramp over the linear ramp
to achieve a more "natural" effect. However, as I experimented
with this and did a little more reading, it seems like that natural
effect really only matters for a recording of natural sounds like
music, speech, or birds. Because we're just pushing around a pure
tone, none of those rules apply and the exponential ramp doesn't
end up solving any of the problems we're experiencing. Instead
the linear ramp is the only way to get the smooth, soft, and pop-less
noise we're looking for. Of course, that comes at the cost of complexity,
but in my opinion it's well worth the health of the patient's ears
and may potentially save gradual damage to headphones over time.

1: https://alemangui.github.io/ramp-to-value