Save session settings in redis
This enables us to both:
a) Pull the full settings from the previous page-view of the
therapist page
b) Pull the current settings from the session for the patient
page

The latter, most importantly, allows us to sync all the settings
when the patient loads the page, including whether to show the
lightbar or the audio indicators.

We store the settings kind of naively in redis using the session
id as the key with json serialized settings.